+++ src/renderer/BlockRenderer.ts
@@ -1,0 +1,3 @@
+export interface BlockRenderer {
+    render(...args: string[]) : string;
+}
