+++ src/renderer/CodeBlockRendered.ts
@@ -1,0 +1,5 @@
+import { BlockRenderer } from "./BlockRenderer";
+
+export interface CodeBlockRenderer extends BlockRenderer {
+    render(block: string, language: string, code: string): string;
+}
